dev-requirements.txt
requests
beautifulsoup4
PyYAML

Flask

python-dotenv
schedule